+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changes by Version
 
 <!-- next version -->
+## v1.27.3
+- Metrics rollup rules:
+  - Support all aggregation types for `MEASUREMENT`
 ## v1.27.2
 - Metrics rollup rules `rollupFunction` field validation for all metric types:
   - `GAUGE`: Any aggregation function supported